Monitoring: Move silence form, types and utils into monitoring/ dir
Clean up of the monitoring code files
  - Move the silence form code to its own file
  - Move the general monitoring types into their own file
  - Move the shared utility functions into a separate file
